+++ src/main/java/com/lowdragmc/photon/client/fx/BlockEffectExecutor.java
@@ -63,7 +63,7 @@ public class BlockEffectExecutor extends FXEffectExecutor {
                 .add(offset.x + 0.5f, offset.y + 0.5f, offset.z + 0.5f));
         root.updateRotation(rotation);
         root.updateScale(scale);
-        this.runtime.emmit(this);
+        this.runtime.emmit(this, delay);
         lastState = level.getBlockState(pos);
         effects.add(this);
     }