+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/data/MaterialSetting.java
@@ -2,10 +2,13 @@ package com.lowdragmc.photon.client.gameobject.emitter.data;
 
 import com.lowdragmc.lowdraglib2.configurator.IConfigurable;
 import com.lowdragmc.lowdraglib2.configurator.annotation.Configurable;
+import com.lowdragmc.lowdraglib2.editor.resource.BuiltinPath;
+import com.lowdragmc.lowdraglib2.editor.resource.IResourcePath;
 import com.lowdragmc.lowdraglib2.syncdata.IPersistedSerializable;
 import com.lowdragmc.photon.client.gameobject.emitter.data.material.BlendMode;
 import com.lowdragmc.photon.client.gameobject.emitter.data.material.IMaterial;
 import com.lowdragmc.photon.client.gameobject.emitter.data.material.TextureMaterial;
+import com.lowdragmc.photon.gui.editor.resource.MaterialResource;
 import com.mojang.blaze3d.systems.RenderSystem;
 import lombok.Getter;
 import lombok.Setter;
@@ -13,6 +16,7 @@ import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 
 import javax.annotation.Nonnull;
+import java.util.Optional;
 
 /**
  * @author KilaBash
@@ -25,7 +29,7 @@ import javax.annotation.Nonnull;
 public class MaterialSetting implements IConfigurable, IPersistedSerializable {
     @Nonnull
     @Configurable(name = "material")
-    protected IMaterial material = new TextureMaterial();
+    protected IMaterial material = Optional.ofNullable(MaterialResource.INSTANCE.getResourceInstance().getResource(new BuiltinPath("circle"))).orElseGet(TextureMaterial::new);
     @Configurable(name = "MaterialSetting.blendMode", subConfigurable = true)
     protected final BlendMode blendMode = new BlendMode();
     @Configurable