+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/data/shape/Box.java
@@ -62,7 +62,7 @@ public class Box implements IShape {
                 pos.y = random.nextFloat() > 0.5 ? scale.y : -scale.y;
             }
         }
-        particle.setLocalPos(Vector3fHelper.rotateYXY(new Vector3f(pos), rotation).add(position).add(particle.getLocalPos()), true);
+        particle.setLocalPos(Vector3fHelper.rotateYXY(new Vector3f(pos), rotation).add(position).add(particle.getLocalPoseWithoutNoise()), true);
         particle.setInternalVelocity(Vector3fHelper.rotateYXY(new Vector3f(0, 0.05f, 0), rotation));
     }