+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/data/shape/Circle.java
@@ -46,7 +46,7 @@ public class Circle implements IShape {
                 0f,
                 (float) (r * Math.sin(theta))).mul(scale);
 
-        particle.setLocalPos(Vector3fHelper.rotateYXY(new Vector3f(pos), rotation).add(position).add(particle.getLocalPos()), true);
+        particle.setLocalPos(Vector3fHelper.rotateYXY(new Vector3f(pos), rotation).add(position).add(particle.getLocalPoseWithoutNoise()), true);
         particle.setInternalVelocity(Vector3fHelper.rotateYXY(new Vector3f(pos).normalize().mul(0.05f), rotation));
     }