+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/data/shape/Dot.java
@@ -10,7 +10,7 @@ public class Dot implements IShape {
 
     @Override
     public void nextPosVel(TileParticle particle, IParticleEmitter emitter, Vector3f position, Vector3f rotation, Vector3f scale) {
-        particle.setLocalPos(position.add(particle.getLocalPos()), true);
+        particle.setLocalPos(position.add(particle.getLocalPoseWithoutNoise()), true);
         particle.setInternalVelocity(new Vector3f(0, 0, 0));
     }
 }