+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/data/shape/Sphere.java
@@ -49,7 +49,7 @@ public class Sphere implements IShape {
         ).mul(scale);
 
 
-        particle.setLocalPos(Vector3fHelper.rotateYXY(new Vector3f(pos), rotation).add(position).add(particle.getLocalPos()), true);
+        particle.setLocalPos(Vector3fHelper.rotateYXY(new Vector3f(pos), rotation).add(position).add(particle.getLocalPoseWithoutNoise()), true);
         particle.setInternalVelocity(Vector3fHelper.rotateYXY(new Vector3f(pos).normalize().mul(0.05f), rotation));
     }